Revert "port of LL renderpipeline/Kirstens S19 pipeline for bridging to Viewer 2 texture system"
This reverts commit 087e15e89930d51c3964329befb273ae3b2d330d.

Conflicts:

	linden/indra/newview/llsurface.cpp
	linden/indra/newview/llviewerwindow.cpp
	linden/indra/newview/llvoavatar.cpp
	linden/indra/newview/pipeline.cpp
	linden/indra/newview/pipeline.h

+++ linden/indra/llmath/llmath.h
@@ -203,7 +203,7 @@ inline S32 llfloor( F32 f )
 		}
 		return result;
 #else
-		return (S32)floorf(f);
+		return (S32)floor(f);
 #endif
 }
